Added image files and add logic for statically putting them on the tiles.
