--- lanterns/src/LakeTile.css
@@ -1,55 +1,0 @@
-.parentContainer {
-    display: block;
-    width: 42px;
-    height: 42px;
-    overflow: hidden;
-    position: relative;
-  }
-  .parent {
-    position: absolute;
-    margin: auto;
-    display: grid;
-    transform: rotate(45deg);  
-    top: -11px;
-    left: -11px;
-  }
-  .topTri {
-    background-color: blue;
-    height: 2rem;
-    width: 2rem;
-    grid-column-start: 1;
-    grid-column-end: 2;
-    grid-row-start: 1;
-    grid-row-end: 2;
-    justify-self: end;
-  }
-  .rightTri {
-    background-color: red;
-    height: 2rem;
-    width: 2rem;
-    grid-column-start: 2;
-    grid-column-end: 3;
-    grid-row-start: 1;
-    grid-row-end: 2;
-  }
-  .bottomTri {
-    background-color: yellow;
-    height: 2rem;
-    width: 2rem;
-    justify-self: end;
-  }
-  .leftTri {
-    background-color: green;
-    height: 2rem;
-    width: 2rem;
-  }
-  .overlayDiv {
-    justify-items: center;
-    position: absolute;
-    height: 1rem;
-    width: 1rem;
-    top: 1rem;
-    left: 50%;
-    height: 1rem;
-    overflow: 
-  }
